added searchForm, date fix
