+++ src/components/News/NewsSearchForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './NewsSearchForm';
